+++ src/lib/datatransfers/magic_wormhole/models.py
@@ -1,0 +1,8 @@
+from lib.datatransfers.datatransfer_base import (
+    WormholeDataTransferDirective,
+    DataTransferOperation,
+)
+
+
+class WormholeDataTransferOperation(DataTransferOperation):
+    transfer_directives: WormholeDataTransferDirective