+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Compress and Extract end-points now support multiple compresion types (none, bz2, gzip, and xz).
+- Support for Magic Wormhole data transfer method
 
 ### Changed
 
 - The System Name path parameter and the corresponding Cluster name configuration are case insensitive.
+- Upload and Download transfer endpoints now require to specify transfer directives
 - Installation docs:
     - Helm charts: FirecREST settings are all included in values.yaml file
     - Changed documentation name from Deployment to Install